removed radius limits
