+++ .gitignore
@@ -46,3 +46,11 @@ docs/build/
 # folders generated by pytest-cov
 htmlcov/
 test-reports/
+
+# jupyter notebook cache directory
+.ipynb_checkpoints
+# jupyter notebook files
+.ipynb
+
+# ppp output directory
+manifest_processed/
