+++ tests/test_parametric_shapes/test_RotateMixedShape.py
