+++ tests/test_parametric_shapes/test_RotateSplineShape.py
