+++ tests/test_parametric_shapes/test_RotateStraightShape.py
